Rename to gameserver_create.js
