+++ shared/gameserver_create.js
